// generated/inputTypeSchemas/AffiliateWhereUniqueInputSchema.ts
import type { Prisma } from '@prisma/client';

import { z } from 'zod';
import { AffiliateWhereInputSchema } from './AffiliateWhereInputSchema';
import { StringFilterSchema } from './StringFilterSchema';
import { DateTimeFilterSchema } from './DateTimeFilterSchema';
import { StringNullableFilterSchema } from './StringNullableFilterSchema';
import { UserScalarRelationFilterSchema } from './UserScalarRelationFilterSchema';
import { UserWhereInputSchema } from './UserWhereInputSchema';
import { AddressNullableScalarRelationFilterSchema } from './AddressNullableScalarRelationFilterSchema';
import { AddressWhereInputSchema } from './AddressWhereInputSchema';
import { OrderListRelationFilterSchema } from './OrderListRelationFilterSchema';

export const AffiliateWhereUniqueInputSchema: z.ZodType<Prisma.AffiliateWhereUniqueInput> = z.object({
  id: z.string().uuid()
})
.and(z.object({
  id: z.string().uuid().optional(),
  AND: z.union([ z.lazy(() => AffiliateWhereInputSchema),z.lazy(() => AffiliateWhereInputSchema).array() ]).optional(),
  OR: z.lazy(() => AffiliateWhereInputSchema).array().optional(),
  NOT: z.union([ z.lazy(() => AffiliateWhereInputSchema),z.lazy(() => AffiliateWhereInputSchema).array() ]).optional(),
  businessName: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  createdAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  updatedAt: z.union([ z.lazy(() => DateTimeFilterSchema),z.coerce.date() ]).optional(),
  userId: z.union([ z.lazy(() => StringFilterSchema),z.string() ]).optional(),
  addressId: z.union([ z.lazy(() => StringNullableFilterSchema),z.string() ]).optional().nullable(),
  user: z.union([ z.lazy(() => UserScalarRelationFilterSchema),z.lazy(() => UserWhereInputSchema) ]).optional(),
  address: z.union([ z.lazy(() => AddressNullableScalarRelationFilterSchema),z.lazy(() => AddressWhereInputSchema) ]).optional().nullable(),
  Order: z.lazy(() => OrderListRelationFilterSchema).optional()
}).strict());

export default AffiliateWhereUniqueInputSchema;